fix:본인만 수정/삭제 가능 및 추천 여부 리턴

+++ server/src/main/java/com/pathfinder/server/diary/service/DiaryService.java
@@ -1,7 +1,10 @@
 package com.pathfinder.server.diary.service;
 
+import com.pathfinder.server.auth.utils.SecurityUtil;
 import com.pathfinder.server.diary.entity.Diary;
 import com.pathfinder.server.diary.repository.DiaryRepository;
+import com.pathfinder.server.global.exception.diaryexception.DiaryDeleteUnAuthorizedException;
+import com.pathfinder.server.global.exception.diaryexception.DiaryEditUnAuthorizedException;
 import com.pathfinder.server.global.exception.diaryexception.DiaryNotFoundException;
 import com.pathfinder.server.member.entity.Member;
 import com.pathfinder.server.member.service.MemberService;
@@ -38,15 +41,20 @@ public class DiaryService {
     @Transactional(propagation = Propagation.REQUIRED, isolation = Isolation.SERIALIZABLE)
     public Diary updateDiary(Diary diary){
         Diary findDiary = findVerifiedDiary(diary.getDiaryId());
-        Optional.ofNullable(diary.getTitle())
-                .ifPresent(title -> findDiary.setTitle(title));
-        Optional.ofNullable(diary.getContent())
-                .ifPresent(content -> findDiary.setContent(content));
-        Optional.ofNullable(diary.getArea1())
-                .ifPresent(area1 -> findDiary.setArea1(area1));
-        Optional.ofNullable(diary.getArea2())
-                .ifPresent(area2 -> findDiary.setArea2(area2));
-        return diaryRepository.save(findDiary);
+        if(verifyIdentification(findDiary)){
+            Optional.ofNullable(diary.getTitle())
+                    .ifPresent(title -> findDiary.setTitle(title));
+            Optional.ofNullable(diary.getContent())
+                    .ifPresent(content -> findDiary.setContent(content));
+            Optional.ofNullable(diary.getArea1())
+                    .ifPresent(area1 -> findDiary.setArea1(area1));
+            Optional.ofNullable(diary.getArea2())
+                    .ifPresent(area2 -> findDiary.setArea2(area2));
+            return diaryRepository.save(findDiary);
+        }
+        else {
+            throw new DiaryEditUnAuthorizedException();
+        }
     }
 
     @Transactional(readOnly = true)
@@ -76,9 +84,14 @@ public class DiaryService {
 
     public void deleteDiary(Long diaryId) {
         Diary findDiary = findVerifiedDiary(diaryId);
-
-        diaryRepository.delete(findDiary);
+        if(verifyIdentification(findDiary)){
+            diaryRepository.delete(findDiary);
+        }
+        else {
+            throw new DiaryDeleteUnAuthorizedException();
+        }
     }
+
 
     @Transactional(readOnly = true)
     public Diary findVerifiedDiary(Long diaryId) {
@@ -91,5 +104,11 @@ public class DiaryService {
         diary.setName(findUser.getName());
         findUser.setDiaryCount(findUser.getDiaryCount() + 1);
         rewardService.unlockRewards(findUser,findUser.getRewards());
+    }
+    public boolean verifyIdentification(Diary diary) {
+        if(diary.getMember().getMemberId() == SecurityUtil.getCurrentId()) {
+            return true;
+        }
+        return false;
     }
 }
